Make module commands selection-aware
Act on the selection or current module, if any, and fallback to
reading a single module instead of acting on all modules without the
option to act only on a subset.  To act on all suitable modules a
prefix argument has to be used now.

Also improve the doc-strings and add comments explaining differences
to Git.